Add exceptions.
This is really just the most sensible way to report certain kinds of errors.

We will probably use them more in DAG operations than AST parsing: in parsing, we want to continue parsing for a "little while" and output useful diagnostic messages if we can. Failures in semantic analysis, conversion to the DAG or backend operations are more fail-stop kind of things.