Require actions to define their parameters.
In addition to arguments that define the action --- e.g. cc = action(command = '$CC') --- we must also define the parameters that an action takes. For instance:

cc = action('$CC $defines -c $in -o $out' <- defines:string)

In addition to being good hygiene, this explicit specification allows us to expose all of the files that should be involved in the dependency graph. For instance, we can pass the --header-file argument to yacc in order to specify where the output header should go; this header should be part of the dependency graph.