Add debug stream handling.
When a caller asks for the debug stream, return either a target Bytestream (Bytestream::Stderr() by default) or Bytestream::None(), depending on the name that is passed to Bytestream::Debug().

For instance, when Fabrique is run with --debug=*, all calls to Debug() will return the real output stream. When run with --debug=parser, Debug("parser.foo") will return the real stream but Debug("dag.bar") will return the do-nothing stream.