Add some Doxygen pieces.
This will be a lot more useful once Fabriquer is self-hosting, as we will be able to define arbitrary Doxygen build steps (e.g. combining with Jekyll) as part of the build. In the meantime, I can't be bothered to write lots of Python that I know will have to be rewritten in Fabriquer later.