Fix more warnings from recent Clang.
Printable needs an explicit copy constructor and assignment operator: it is deprecated to use implicitly-generated ones with an explicit destructor (and it's not recommended to use a non-virtual destructor when there are virtual methods).

flex/byacc generate code with the 'register' keyword, but this is now deprecated, so mask this warning on all generated code.