Add support for importing modules.
We can import modules as follows:

foo = import('foo.fab');

This commit enables the syntax, as well as actually parsing the module
in order to determine its (structural) type.  It doesn't include the
ability to actually *use* the imported module, however: that requires
field-access syntax, which is next on the agenda.