Move detailed argument checks from Parsing to DAG.
At the point of parsing a function call, we might not actually know things like the parameter names: we might only know that we have a reference to a function of a certain type.

We can still do a little sanity checking during parsing, but not detailed evaluation of argument types (or even number: an action’s output looks syntactically like parameters).