Move builtin generation out of EnterScope().
Rather than giving EnterScope() magic functionality depending on whether or not we’re already in a scope, make the driver explicitly create the top-level builtins at the top level and provide the mechanism for additional builtins, should we need them.

Now we don’t need a buildroot_ member in Parser.