Refactor function calls to use callbacks.
The previous function-call code introduced an AST/DAG layering violation: a dag::Function needed a reference to an ast::Function in order to do its work. Now, we can define arbitrary callbacks to handle function calls. This will be extra nice when we start writing plugins that expose non-AST-based functions!