Allow overriding file-not-found behaviour.
Rather than always throwing an exception, let FindFile() take a std::function to call when we want to report a missing file. This can default to fabrique::FileNotFound(), which does throw an exception, but other code (like FindModule()) might want to simply return an empty string because not finding a module in a particular set of paths is quite normal.