Remove regeneration from bootstrap makefile.
Depending on when various files were created by the revision control
client, make might first want to re-build the makefile, which requires a
working Fabrique instance. We don't have one of those yet: that's why
we're bootstrapping!

Solve the problem by removing the regeneration facility from the bootstrap
makefile. If we need to regenerate this file, we need to do it explicitly.