Move output directory creation into DAGBuilder.
There are times when the build system needs to know how to generate output directories (e.g., when a build depends on a set of generated include directories). We used to handle this as a special case in the Make backend, but we should handle it with other backends too, as the include-directory case applies everywhere.