Remove requirement for AST nodes to have types.
There will be lots of circumstances in which the types of our AST nodes won't be known until run time. Relax the requirement for AST Expressions to have types.

This breaks parsing, but I won't bother to fix the old yacc-based parser since we're going to build a new one.