Add support for arithmetic negation operation.
"Negate" isn't the right name for the logical `not` operator: that's a
better name for arithmetic negation (`-`). Rename existing `Negate` to
`Not` and add a new, more appropriate `Negate` operation.

While we're here, remove `dag::Build::Negate`, as it doesn't make sense
to negate (or apply `not` to) the results of a build, which are files.