Improve/clarify/simplify lexical scope management.
Distinguish between lexical scopes involved in AST node evaluation
(which cannot, by definition, live longer than the evaluation) and
groups of values that were defined together in a lexical scope (which
can live on). This simplifies the logic of AST node evaluation slightly
and lays the foundation for passing sets of values around for things
like function calls (which can capture values from the surrounding
lexical scope).

Also, remove the requirement to call scope.leave() on leaving a scope:
let that be done by a destructor to ensure we don't forget it.