Fix value capturing for lambda functions.
Extend `EvalContext::EnterScope()`, allowing the parent scope to be
specified explicitly. This allows us to specify that, when a lookup in
the current scope fails, further lookups should happen in the scope of
the function definition (which is preserved via a `shared_ptr` to an
`EvalContext::ScopedValues` object) rather than the call.