Refactor builtin creation.
Extract the generation of builtins to a separate source file and allow
builtins to be generated after the EvalContext has been set up. This
helps with the chicken-and-egg problem of needing an EvalContext to
implement the `import()` builtin but needing to pass builtins to the
EvalContext when it is created.