Fix list types (properly! finally!).
Simplify SequenceType by allowing it to either have a parameter (for a
list of things) or not (for an empty list). Let an unparameterized
SequenceType be a subtype of a parameterized one, allowing us to pass an
empty list (of type `list`) to a parameter of type `list[foo]`.
This allows us to remove the RawSequenceType, clean up some places where
we previously needed `dynamic_cast<>`, etc.