Remove [some] hardcoded directory separators.
Now that we have a sensible file prefix operator, use it! Remove places
where we were generating directory+filenames in our build description
with hardcoded directory separators, other than within `import()`.
We will be able to sort that out too once `import()` accepts a `file`
as an argument instead of a `string`.