Ignore unknown compiler warning options.
In the future we should do more to introspect on the version of Clang
being run and figure out what warnings it really supports. Or,
alternatively, we could look for a Fabrique configuration file to ship
with Clang that describes various categories of warnings. For now,
however, just ignore the `-Wno-foo` options that the version of the
compiler we're executing doesn't understand.