Rework Fabrique::Parse() to allow absolute paths.
We would like to be able to process a Fabrique file using its absolute
path, and we would also like to not calculate that path multiple times.
To attain both of these goals, reword `Fabrique::Parse()` to take a
`std::istream&` as well as a filename and move some of the "is this a
file" checks to `Fabrique::Process()`. Thus, we can emit "no such file"
messages using the filename that the user typed in while passing the
absolute path into the `Parser` to keep our input file paths explicit.